crf masks change datasets.py file to output val teacher masks

+++ models_mage_2dec.py
@@ -853,7 +853,9 @@ class MaskedGenerativeEncoderViT(nn.Module):
 
         
         with torch.cuda.amp.autocast(enabled=False):
-            mask_crf_onehot = torch.nn.functional.one_hot(mask_crf, num_classes=self.slot_attention.num_slots).permute(0,3,1,2).to(dtype=torch.float16)
+            # mask_crf_onehot = torch.nn.functional.one_hot(mask_crf, num_classes=self.slot_attention.num_slots).permute(0,3,1,2).to(dtype=torch.float16)
+            mask_crf_onehot = torch.nn.functional.one_hot(mask_crf, num_classes=self.slot_attention.num_slots).permute(0,3,1,2).to(dtype=torch.float)
+
             interpolated = F.interpolate(mask_crf_onehot, size=(16, 16), mode='nearest')
             mask_crf_slots = interpolated.view(interpolated.shape[0], interpolated.shape[1], -1)
